backend/requirements.txt
fastapi
uvicorn[standard]
boto3
openai
python-multipart
SQLModel
sqlalchemy
psycopg2-binary
python-dotenv
aiofiles
passlib[bcrypt]
python-jose[cryptography]
pydub
python-magic-bin
